Refactor WebVoyager benchmark structure: update imports, add commands and dataset files, implement runner logic, and remove outdated tests.
